[#12][MODIFY] 일기&데일리 질문 패키지 병합
