+++ src/main/java/com/kancho/byeolbyeol/daily/domain/daily_question/DailyQuestionRepository.java
@@ -1,0 +1,10 @@
+package com.kancho.byeolbyeol.daily.domain.daily_question;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.time.LocalDate;
+import java.util.Optional;
+
+public interface DailyQuestionRepository extends JpaRepository<DailyQuestion, Long> {
+    Optional<DailyQuestion> findByDate(LocalDate nowLocalDate);
+}
