+++ src/main/java/com/kancho/byeolbyeol/daily/dto/ReqWriteDiaryDto.java
@@ -1,0 +1,22 @@
+package com.kancho.byeolbyeol.daily.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReqWriteDiaryDto {
+
+    @NotNull
+    private String title;
+
+    @NotNull
+    private String content;
+
+    @NotNull
+    private Long horoscopeId;
+}
