--- src/main/java/com/kancho/byeolbyeol/diary/domain/DiaryRepository.java
@@ -1,14 +1,0 @@
-package com.kancho.byeolbyeol.diary.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import javax.transaction.Transactional;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.Date;
-import java.util.Optional;
-
-public interface DiaryRepository extends JpaRepository<Diary, Long> {
-    Optional<Diary> findByUsersIdAndDateBetween(Long userId, LocalDateTime startTime, LocalDateTime endTime);
-}
