--- src/main/java/com/kancho/byeolbyeol/diary/dto/ReqModifyDiaryDto.java
@@ -1,19 +1,0 @@
-package com.kancho.byeolbyeol.diary.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ReqModifyDiaryDto {
-
-    @NotNull
-    private String title;
-
-    @NotNull
-    private String content;
-}
