--- src/main/java/com/kancho/byeolbyeol/diary/exception/NotFoundDiaryException.java
@@ -1,24 +1,0 @@
-package com.kancho.byeolbyeol.diary.exception;
-
-import com.kancho.byeolbyeol.exception.BaseException;
-import com.kancho.byeolbyeol.exception.ErrorModel;
-import org.springframework.http.HttpStatus;
-
-public class NotFoundDiaryException extends BaseException {
-    public NotFoundDiaryException() {
-        this(HttpStatus.BAD_REQUEST);
-    }
-
-    private NotFoundDiaryException(HttpStatus httpStatus) {
-        this(4010, httpStatus);
-    }
-
-    private NotFoundDiaryException(int code, HttpStatus httpStatus) {
-        super(ErrorModel.builder()
-                .code(code)
-                .httpStatus(httpStatus)
-                .massage("Not Found Diary")
-                .build());
-    }
-}
-
